test(exec-batch): split perScriptTimeoutMs tests into sibling file and add parallel-path regression
Move the two perScriptTimeoutMs regression tests (#77) from exec-batch.test.ts
into exec-batch.perscript.test.ts to bring exec-batch.test.ts back under the
300-line guideline. Add a third test exercising the runParallel code path
(readOnly:true) to cover the structurally different per-script AbortController
mechanism used in that branch — the parallel regression was missing from review.
