+++ src/api/tests/unit/exec-batch.test.ts
@@ -268,147 +268,4 @@ describe("POST /v1/sandboxes/:id/exec-sync-batch", () => {
 
 		expect(res.status).toBe(404);
 	});
-
-	// Regression for issue #77 — shared timeout silently kills later scripts.
-	// Before the fix: totalTimeoutMs was the only budget, so a slow first script
-	// would exhaust the shared budget and later scripts came back exit_code=-1,
-	// error="timeout" even though they would have succeeded individually.
-	// After the fix: perScriptTimeoutMs gives each script its own independent budget.
-	it("perScriptTimeoutMs: later scripts succeed when first script times out (regression #77)", async () => {
-		vi.useFakeTimers();
-		try {
-			const { sessionManager } = await makeTestEnv();
-			const app = makeTestApp(sessionManager);
-			const token = await makeToken();
-
-			const batchSandboxId = `${SANDBOX_ID}-per-script`;
-			const session = await sessionManager.getOrCreate("default", batchSandboxId);
-
-			// Latch fires once the first blocking exec is registered so we can advance
-			// the fake clock after its per-script timer has been created.
-			let latchResolve!: () => void;
-			const execBlocking = new Promise<void>((r) => {
-				latchResolve = r;
-			});
-
-			let callCount = 0;
-			vi.spyOn(session.bash, "exec").mockImplementation((_script, opts) => {
-				callCount++;
-				if (callCount === 1) {
-					// First script blocks forever — simulates Python 1.4s WASM startup.
-					return new Promise((_resolve, reject) => {
-						latchResolve();
-						opts?.signal?.addEventListener("abort", () => {
-							reject(new DOMException("The operation was aborted", "AbortError"));
-						});
-					});
-				}
-				// Subsequent scripts succeed instantly.
-				return Promise.resolve({ stdout: "ok\n", stderr: "", exitCode: 0, env: {} });
-			});
-
-			const resPromise = app.request(`/v1/sandboxes/${batchSandboxId}/exec-sync-batch`, {
-				method: "POST",
-				headers: { Authorization: `Bearer ${token}`, "Content-Type": "application/json" },
-				body: JSON.stringify({
-					scripts: [
-						{ id: "slow", script: "sleep 1000" },
-						{ id: "fast1", script: "echo ok" },
-						{ id: "fast2", script: "echo ok" },
-					],
-					// Total budget is 5s — far more than enough for fast1 and fast2.
-					// Per-script budget is 50ms — just enough to kill the first slow script.
-					// Before the fix (no perScriptTimeoutMs wiring): all three would share
-					// the 5s total, and the first script would block for the full 5s, causing
-					// fast1 and fast2 to come back as timeout false-negatives.
-					timeoutMs: 5000,
-					perScriptTimeoutMs: 50,
-				}),
-			});
-
-			// Wait until the mock is blocking before advancing the clock.
-			await execBlocking;
-			// Advance past the per-script timeout for the first script only (50ms).
-			await vi.advanceTimersByTimeAsync(60);
-
-			const res = await resPromise;
-			expect(res.status).toBe(200);
-			const body = (await res.json()) as { results: BatchResult[] };
-
-			// The first (slow) script must be killed by its per-script timeout.
-			expect(body.results[0]!.id).toBe("slow");
-			expect(body.results[0]!.exitCode).toBe(-1);
-			expect(body.results[0]!.error).toBe("timeout");
-
-			// The later scripts must NOT be false-negatives — they succeed independently.
-			// Before the fix these would both be exit_code=-1, error="timeout".
-			expect(body.results[1]!.id).toBe("fast1");
-			expect(body.results[1]!.exitCode).toBe(0);
-			expect(body.results[1]!.error).toBeUndefined();
-
-			expect(body.results[2]!.id).toBe("fast2");
-			expect(body.results[2]!.exitCode).toBe(0);
-			expect(body.results[2]!.error).toBeUndefined();
-		} finally {
-			vi.useRealTimers();
-		}
-	});
-
-	it("perScriptTimeoutMs is capped by totalTimeoutMs outer ceiling", async () => {
-		vi.useFakeTimers();
-		try {
-			const { sessionManager } = await makeTestEnv();
-			const app = makeTestApp(sessionManager);
-			const token = await makeToken();
-
-			const batchSandboxId = `${SANDBOX_ID}-ceiling`;
-			const session = await sessionManager.getOrCreate("default", batchSandboxId);
-
-			let latchResolve!: () => void;
-			const execBlocking = new Promise<void>((r) => {
-				latchResolve = r;
-			});
-
-			vi.spyOn(session.bash, "exec").mockImplementation((_script, opts) => {
-				// Every script blocks forever — only abort can resolve.
-				return new Promise((_resolve, reject) => {
-					latchResolve();
-					latchResolve = () => {}; // prevent double-resolve
-					opts?.signal?.addEventListener("abort", () => {
-						reject(new DOMException("The operation was aborted", "AbortError"));
-					});
-				});
-			});
-
-			const resPromise = app.request(`/v1/sandboxes/${batchSandboxId}/exec-sync-batch`, {
-				method: "POST",
-				headers: { Authorization: `Bearer ${token}`, "Content-Type": "application/json" },
-				body: JSON.stringify({
-					scripts: [
-						{ id: "a", script: "block" },
-						{ id: "b", script: "block" },
-					],
-					// Total budget 100ms is smaller than per-script budget 10000ms.
-					// The total deadline must fire first and kill everything.
-					timeoutMs: 100,
-					perScriptTimeoutMs: 10000,
-				}),
-			});
-
-			await execBlocking;
-			await vi.advanceTimersByTimeAsync(110);
-
-			const res = await resPromise;
-			expect(res.status).toBe(200);
-			const body = (await res.json()) as { results: BatchResult[] };
-
-			// Both scripts timed out — the outer ceiling took effect.
-			for (const result of body.results) {
-				expect(result.exitCode).toBe(-1);
-				expect(result.error).toBe("timeout");
-			}
-		} finally {
-			vi.useRealTimers();
-		}
-	});
 });
